Create x's and o's
This function takes a string and determines whether the number of x's and o's is equal, returning "true" if it is and "false" if it isn't. The function compares the "xCount" and "oCount", and consideration must be given to lowercase the letters in the string to receive accurate results. If there are no x's and o's in the string, then the result is also true.